Squashed commit of the following:
commit c720e6d95cd5c3a18ad979159b5584aa63395ffb
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Mon Dec 23 12:41:31 2024 -0600

    Local documentation build conveniences (#4592)

    * A local build Makefile

    * add more stuff for local build to environment.yml

    * Add hobu logo stuff

    * remove conf.py – this is now generated by jupyterbook

    * drop conda-provided sed/graphviz/doxygen

    * a little doc for building locally

    * update substitutions

commit 74382cbeda4b2dd60bbcc021f410ac74392e94d2
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Mon Dec 23 09:52:32 2024 -0600

    adapt standard gitignore for CMake files (#4596)

commit 9bab3a3043abc59c498d383ba9f241bca9f36196
Author: Guilhem Villemin <guilhem.villemin@altametris.com>
Date:   Tue Dec 17 18:13:22 2024 +0100

    Enhance `filters.gpstimeconvert` to make it streamable (#4568)

    * make filter streamable

    * fix typo in disclaimer

    * update doc

    * thow when invalid conversion

    * revert spurious edit

    * add missing Streamable

    * adding run method

    * add wrapped tolerance arg

    ---------


commit 991e67c099bbdf2462f35bfe154778121938504a
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Tue Dec 17 09:51:10 2024 -0600

    Fix #4578, doc nit in readers.rdb (#4590)

commit 998be563b8193e97798eb9ed8334b8cf4827a383
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Mon Dec 16 10:30:14 2024 -0600

    update community page with link to Matrix (#4587)

commit 2d4a3cea0188664df459022e574d01a2005f2d9e
Author: Andrew Bell <andrew.bell.ia@gmail.com>
Date:   Mon Dec 16 09:47:35 2024 -0500

    Make sure points are added to a new point view if a filter returns no views and (#4583)

    where_merge == 'true'.
    Close #4572

commit cda06317f7221d486208423e3e51e9122aad7c0d
Author: Andrew Bell <andrew.bell.ia@gmail.com>
Date:   Mon Dec 16 09:39:22 2024 -0500

    Write binary in binary mode. (#4584)

    Some general cleanup.

commit 213a6a631f57ba6cd38d384652958264b74e5501
Author: chambbj <brad.chambers@gmail.com>
Date:   Thu Dec 12 12:13:34 2024 -0600

    Isolate dependencies and tasks for doc generation and preview (#4579)

commit 2271196311b2ccb05d8b61c3852925cca7b333ca
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Thu Dec 12 11:59:06 2024 -0600

    adapt GDAL's backport bot (#4580)

commit 97522d1d868f1509504408f9c60e00247290993a
Author: chambbj <brad.chambers@gmail.com>
Date:   Wed Dec 11 11:07:00 2024 -0600

    Use tabs for the OS variants of using Conda (#4575)

commit ab8a5d18b39271049b4ed05af3f4acc2968be223
Author: chambbj <brad.chambers@gmail.com>
Date:   Wed Dec 11 10:45:39 2024 -0600

    Rewicker the pixi config (#4571)

    * Rewicker the pixi config

    * Enable NITF and also mark this as a win-64 platform task

    * Use target_compile_definitions for the _XKEYCHECK_H fix

    * Add task for linux-64 configure, move docs, build, and test tasks

commit 88531fc59116fd9d7b18ae3e26e6c8f46c2853db
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Wed Dec 4 17:16:29 2024 -0600

    protect against including null strings when reading PDAL vlr types (#4566)

commit 980095719e6beeb9321713719b87527698e46791
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Wed Dec 4 15:23:11 2024 -0600

    bump master CMake version to 2.9.0 to eliminate confusion

commit d172b323f37ff18e68248d5493a3b1589a9543b1
Author: chambbj <brad.chambers@gmail.com>
Date:   Wed Dec 4 15:11:26 2024 -0600

    Add Pixi install instructions (#4565)

    * Add instructions for setting up a Pixi environment

    * Fix nested code block

commit de41500d14bc02d8983c81d43b8e0a62ba477c44
Author: Andrew Bell <andrew.bell.ia@gmail.com>
Date:   Tue Dec 3 15:58:31 2024 -0500

    `filters.expression` support for filtering NaN and LOWEST/HIGHEST (#4561)

    * Checkpoint.

    * testing WIP

    * Checkpoint.

    * Don't call checkEnd() in MathParser.
    Add failure tests.

    * Remove dead comment.

    * Provide nan support.

    * Remove junk.

    * Fix typos.

    * Test.

    * Test.

    * Test.

    * Test 2.

    * Test 3.

    * Test 4.

    * Test 5.

    * bump ci

    ---------


commit cc3e18ff09a0b35c9c2b82448cbe4a00c09e8e14
Author: Guilhem Villemin <guilhem.villemin@altametris.com>
Date:   Tue Dec 3 17:28:43 2024 +0100

    Extend `filters.gpstimeconvert` to handle day seconds.

    * convert to day seconds

    * typo

    * finish day conversion

    * fixes

    * restore conversion arg, mark it as deprecated

    * declare m_converison

    * fix type declaration

commit cf6f6ef95575359d0536ab5cc7225cfd1101c5a1
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Sun Dec 1 10:37:20 2024 -0600

    Validate checksums and version number in automated release CI (#4564)

    * add release version checks

    * set exit

    * dummy 7.7.7 release version

    * explicit verify of checksums

    * cd into correct dir

    * need conda for testing

    * syntax

    * break apart steps a bit

    * set path to environment

    * bad path :(

    * bad path again :(

    * dont build tests

    * check installed version version

    * use pdal-config for clean version string

    * only release for x.x.x

commit 3334e6ecd666b2ff8bc7c427bf7864a3e8ba64f1
Author: chambbj <brad.chambers@gmail.com>
Date:   Fri Nov 22 15:52:28 2024 -0600

    Detect path to conf.py and set JSON/CSV paths relative (#4560)

    * Detect path to conf.py and set JSON/CSV paths relative

    * Try running it as a separate pre_build step

    * fix typo

    * jeez

    ---------


commit eab55c5e4e1d4a8d6745041b5d14efbf0985be30
Author: Howard Butler <hobu.inc@gmail.com>
Date:   Wed Nov 20 16:18:38 2024 -0600

    Automate release creation (#4559)

    * automate release artifact generation

    * on push

    * fix syntax

    * test builds for any tag

    * explicit source tree build job

    * working-directory

    * only draft release for x.x.x tags

    * fix syntax

    * update HOWTORELEASE

+++ test/unit/io/LasWriterTest.cpp
@@ -2293,7 +2293,7 @@ TEST(LasWriterTest, issue2235)
         writer41.execute(table41);
 
         const las::Header& h41 = tester.header(writer41);
-        EXPECT_EQ(h41.legacyPointCount, (int32_t)h41.pointCount());
+        EXPECT_EQ(h41.legacyPointCount, h41.pointCount());
 
         for (int i=0; i<h41.LegacyReturnCount; i++)
             EXPECT_EQ(h41.legacyPointsByReturn[i], h41.ePointsByReturn[i]);
@@ -2316,14 +2316,14 @@ TEST(LasWriterTest, issue2235)
 
         const las::Header& h46 = tester.header(writer46);
 
-        EXPECT_EQ(h46.legacyPointCount, 0);
-        EXPECT_NE(h46.pointCount(), 0);
+        EXPECT_EQ(h46.legacyPointCount, 0UL);
+        EXPECT_NE(h46.pointCount(), 0UL);
 
         for (int i=0; i<h46.LegacyReturnCount; i++)
         {
-            EXPECT_EQ(h46.legacyPointsByReturn[i], 0);
+            EXPECT_EQ(h46.legacyPointsByReturn[i], 0UL);
             if (i < returnsWithPoints)
-                EXPECT_NE(h46.ePointsByReturn[i], 0);
+                EXPECT_NE(h46.ePointsByReturn[i], 0UL);
         }
     }
 
@@ -2351,14 +2351,14 @@ TEST(LasWriterTest, issue2235)
 
         const las::Header& headerCopc = tester.header(readerCopc);
 
-        EXPECT_EQ(headerCopc.legacyPointCount, 0);
-        EXPECT_NE(headerCopc.pointCount(), 0);
+        EXPECT_EQ(headerCopc.legacyPointCount, 0UL);
+        EXPECT_NE(headerCopc.pointCount(), 0UL);
 
         for (int i=0; i<headerCopc.LegacyReturnCount; i++)
         {
-            EXPECT_EQ(headerCopc.legacyPointsByReturn[i], 0);
+            EXPECT_EQ(headerCopc.legacyPointsByReturn[i], 0UL);
             if (i < returnsWithPoints)
-                EXPECT_NE(headerCopc.ePointsByReturn[i], 0);
+                EXPECT_NE(headerCopc.ePointsByReturn[i], 0UL);
         }
     }
 }
